API and image optimizations
Moving API calls to App.vue and getting images from url instead of assets folder
