+++ src/App.vue
@@ -16,7 +16,7 @@
               </ul>
             </div>
         </nav>
-        <router-view :time="times"></router-view>
+        <router-view :time="times" :partyppl="partypplAPIdata" :bitems="budgetAPIdata" :vendors="vendorsAPIdata" :tasks="tasksAPIdata" :guests="guestAPIdata"></router-view>
         <v-btn v-scroll="onScroll" v-show="fab" fab dark fixed bottom right color="info" @click="toTop">
           <v-icon>mdi-chevron-up</v-icon>
         </v-btn>          
@@ -25,9 +25,16 @@
 </template>
 
 <script>
+import axios from "axios";
 export default {
   data() {
     return {
+    partypplAPIdata: [],
+    budgetAPIdata: [],
+    guestsAPIdata: [],
+    tasksAPIdata: [],
+    vendorsAPIdata: [],
+
     fab: false,
     startTime: Date.now(),
     endTime: "Sept 18, 2021 17:00:00",
@@ -38,9 +45,32 @@ export default {
       { id: 3, text: "Seconds", time: 1 }
     ],
     progress: 100,
-    // isActive: false,
     timeinterval: undefined
     }
+  },
+  mounted() {
+        let sheetsuParty = "https://sheetsu.com/apis/v1.0su/eded8760576f"
+        let sheetsuBudget = "https://sheetsu.com/apis/v1.0su/de90e9842a02"
+        let sheetsuGuests = "https://sheetsu.com/apis/v1.0su/33fd43026ba4"
+        let sheetsuTasks = "https://sheetsu.com/apis/v1.0su/971084032f1d"
+        let sheetsuVendors = "https://sheetsu.com/apis/v1.0su/02e39cefb1c3"
+
+        const requestParty = axios.get(sheetsuParty);
+        const requestBudget = axios.get(sheetsuBudget);
+        const requestGuests = axios.get(sheetsuGuests);
+        const requestTasks = axios.get(sheetsuTasks);
+        const requestVendors = axios.get(sheetsuVendors);
+
+        axios.all([requestParty,requestBudget,requestGuests,requestTasks,requestVendors
+        ]).then(axios.spread((...responses) => {
+            this.partypplAPIdata = responses[0].data
+            this.budgetAPIdata = responses[1].data
+            this.guestAPIdata = responses[2].data
+            this.tasksAPIdata = responses[3].data
+            this.vendorsAPIdata = responses[4].data
+        })).catch(errors => {
+            console.log('Request failed', errors);
+        })
   },
   methods: {
     onScroll (e) {
@@ -62,7 +92,6 @@ export default {
         this.updateProgressBar();
       } else {
         clearTimeout(this.timeinterval);
-        // this.times[3].time = this.times[2].time = this.times[1].time = this.times[0].time = 0;
          this.progress = 0;
       }
     },
@@ -107,7 +136,7 @@ export default {
 }
 
 .main, .details, .rsvp {
-  background-image: url("assets/venue1.jpg");
+  background-image: url("https://i.imgur.com/8k6qUXq.jpg");
   background-position: center center;
   background-repeat: no-repeat;
   background-attachment: fixed;
@@ -145,7 +174,7 @@ li {
 }
 
 .gcard:hover {
-  background-image: url("assets/marbles.jpg");
+  background-image: url("https://i.imgur.com/Uur66PW.jpg");
   background-position: center center;
   background-repeat: no-repeat;
   background-attachment: scroll;
@@ -153,7 +182,7 @@ li {
 }
 
 .bcard:hover {
-  background-image: url("assets/flowers.jpg");
+  background-image: url("https://i.imgur.com/rRBTMqM.jpg");
   background-position: center center;
   background-repeat: no-repeat;
   background-attachment: scroll;
