+++ src/components/main.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="main">
     <div class="container greyblock">
-      <h1>Kevin &amp; Chastity</h1>
+      <h1>Kevin &amp; Chasity</h1>
       <h3>September 18, 2021 in Lahaina, HI</h3>
       <section class="flex-container timer">
           <div class="flex-item number">{{ time[0].time }}</div>
